+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 -   Gave edition and chapter editors permission to upload images.
 -   Make unpublished editions visible to users with edition edit privileges.
 -   Maximum width on select widgets.
+-   Improved position of reference ID in edit mode.
 
 # 0.5.63 - 2023-11-11